Refactor audio and file system for stream-based access
Reworked audio managers to dispose streams after loading and updated file system interfaces to support stream-based access instead of byte arrays. Added SeekableStream for efficient random access, improved memory usage with weak references, and modernized code with collection expressions. Updated project files and solution format for Windows-specific targets and added WinRTMediaPlayer support for desktop audio playback.

+++ ReOsuStoryboardPlayer.Avalonia/Utils/SimpleFileSystem/Impl/AvaloniaStorageProvider/AvaloniaStorageProviderFileSystemBuilder.cs
@@ -16,7 +16,7 @@ public class AvaloniaStorageProviderFileSystemBuilder
                     case IStorageFile childFile:
                         var prop = await childFile.GetBasicPropertiesAsync();
                         var childJsfsFile =
-                            new AvaloniaStorageProviderSimpleFile(jsfsDir, childFile.Name, (int) (prop.Size ?? 0), childFile);
+                            new AvaloniaStorageProviderSimpleFile(jsfsDir, childFile.Name, (long) (prop.Size ?? 0), childFile);
                         jsfsDir.AddFile(childJsfsFile);
                         break;
                     case IStorageFolder childFolder:
